fix avatar cropping on mobile
